Missing docstring.

Missing docstring for RetentionData. Check Documenter's build log for details.

Module Index

RetentionData.ABC

RetentionData.ABC_to_Kcentric

RetentionData.ABC_to_TD

RetentionData.AcceptTest

RetentionData.Kcentric

RetentionData.Kcentric_to_ABC

RetentionData.TD_to_ABC

RetentionData.T_column_names_to_Float

RetentionData.add_group_to_Cat!

RetentionData.align_categories

RetentionData.all_parameters

RetentionData.all_parameters!

RetentionData.coeff_of_determination

RetentionData.collect_categories

RetentionData.collect_csv_paths

RetentionData.dataframe_of_all

RetentionData.duplicated_data

RetentionData.extract_meta_data

RetentionData.extract_parameters_from_fit

RetentionData.fit

RetentionData.fit

RetentionData.fit_models

RetentionData.fit_models

RetentionData.fit_models!

RetentionData.fit_models!

RetentionData.flag

RetentionData.flagged_data

RetentionData.formula_to_dict

RetentionData.indentify_parameters

RetentionData.lambertw_x

RetentionData.load_allparameter_data

RetentionData.load_csv_data

RetentionData.load_csv_data!

RetentionData.load_custom_CI_database

RetentionData.load_lnkT_data

RetentionData.load_parameter_data

RetentionData.new_database_format

RetentionData.old_database_format

RetentionData.plot_lnk_fit

RetentionData.plot_res_lnk_fit

RetentionData.ring_number

RetentionData.save_all_parameter_data

RetentionData.substance_identification

Detailed API

RetentionData.ABC — Method

ABC(x, p)

The ABC-model for the relationship lnk(T).

Arguments:

x: temperature T in K

p[1] = A - lnβ

p[2] = B in K

p[3] = C

RetentionData.ABC_to_Kcentric — Method

ABC_to_Kcentric(A, B, C, KRef)

Convert the parameters of the ABC set (A, B, C) to the parameters of the K-centric set (Tchar, θchar, ΔCp). The reference distribution coefficient KRef is a needed additional information (normally, it should be equal to the phase ratio β of the reference column, where the parameters where measured). The parameter should have been measured in the following units:

A (without unit)

B in K

C (without unit)

KRef (without unit)

Output

Tchar in °C

θchar in °C

ΔCp in J mol⁻¹ K⁻¹

RetentionData.ABC_to_TD — Method

ABC_to_TD(A, B, C, Tref)

Convert the parameters of the ABC set (A, B, C) to the parameters of the thermodynamic set (ΔHref, ΔSref, ΔCp). The reference temperature Tref is a needed additional information. The parameter should have been measured in the following units:

A (without unit)

B in K

C (without unit)

Tref in °C

Output

ΔHref in J mol⁻¹

ΔSref in J mol⁻¹ K⁻¹

ΔCp in J mol⁻¹ K⁻¹

RetentionData.AcceptTest — Method

AcceptTest(Dataset, Check; ftrusted=0.7)

Make the fit of lnk over T for a all solutes of a Dataset with the Kcentric-model. With the boolean variable Check a test for outliers, using a robust fit (RAFF.jl), is executed.

Output

The fit results and additional statistics are exported as a DataFrame. Categories, which are included in the measured lnk-data are appended to the DataFrame.

RetentionData.Kcentric — Method

Kcentric(x, p)

The K-centric model for the relationship lnk(T).

Arguments

x: temperature T in K

p[1] = Tchar + Tst in K

p[2] = θchar in °C

p[3] = ΔCp/R

RetentionData.Kcentric_to_ABC — Method

Kcentric_to_ABC(Tchar, θchar, ΔCp, KRef)

Convert the parameters of the K-centric set (Tchar, θchar, ΔCp) to the parameters of the ABC set (A, B, C). The reference distribution coefficient KRef is a needed additional information (normally, it should be equal to the phase ratio β of the reference column, where the parameters where measured). The parameter should have been measured in the following units:

Tchar in °C

θchar in °C

ΔCp in J mol⁻¹ K⁻¹

KRef (without unit)

Output

A (without unit)

B in K

C (without unit)

RetentionData.TD_to_ABC — Method

TD_to_ABC(ΔHref, ΔSref, ΔCp, Tref)

Convert the parameters of the thermodynamic set (ΔHref, ΔSref, ΔCp) to the parameters of the ABC set (A, B, C). The reference temperature Tref is a needed additional information. The parameter should have been measured in the following units:

ΔHref in J mol⁻¹

ΔSref in J mol⁻¹ K⁻¹

ΔCp in J mol⁻¹ K⁻¹

Tref in °C

Output

A (without unit)

B in K

C (without unit)

RetentionData.T_column_names_to_Float — Method

T_column_names_to_Float(data::DataFrame)

Translates the column names of the dataframe data containing the values of isothermal temperature to Float64. For the case of identical temperatures, a number is appended, separated by an underscore _. If this is the case the string is splited at the _ and only the first part is used.

RetentionData.add_group_to_Cat! — Method

add_group_to_Cat!(newdata)

Add categories defined by the file groups.csv (group name and a list of CAS numbers)

RetentionData.align_categories — Method

align_categories(newdata)

Aligns the entrys for categorys for the same substances (same CAS number).

RetentionData.all_parameters! — Method

all_parameters!(meta_data)

Calculate the other parameter sets from the given parameter sets and replace the data in meta_data with the following:

A new array of dataframes. The dataframes have the following columns:

Name: Name of the substance

A: parameter A of the ABC set

B: parameter B of the ABC set

C: parameter B of the ABC set

Tchar: parameter Tchar of the K-centric set

thetachar: parameter θchar of the K-centric set

DeltaCp: parameter ΔCp of the K-centric and TD set

DeltaHref: parameter ΔHref of the TD set

DeltaSref: parameter ΔSref of the TD set

Tref: the reference temperature of the TD set

beta0: the phase ratio β0, for which the retention was measured, KRef=β0 of the K-centric set

lambertw_x: the argument of the Lambert W function used in the conversion of the ABC set to the K-centric set. The value should be -1/e ≤ x < 0

RetentionData.all_parameters — Method

all_parameters(data::Array{DataFrame,1}, paramset::Array{Array{String,1},1}, paramindex::Array{Dict,1})

Calculate the other parameter sets from the given parameter sets and return them in a new array of dataframes.

Output

A new array of dataframes. The dataframes have the following columns:

Name: Name of the substance

A: parameter A of the ABC set

B: parameter B of the ABC set

C: parameter B of the ABC set

Tchar: parameter Tchar of the K-centric set

thetachar: parameter θchar of the K-centric set

DeltaCp: parameter ΔCp of the K-centric and TD set

DeltaHref: parameter ΔHref of the TD set

DeltaSref: parameter ΔSref of the TD set

Tref: the reference temperature of the TD set

beta0: the phase ratio β0, for which the retention was measured, KRef=β0 of the K-centric set

lambertw_x: the argument of the Lambert W function used in the conversion of the ABC set to the K-centric set. The value should be -1/e ≤ x < 0

RetentionData.coeff_of_determination — Method

coeff_of_determination(fit, y)

Calculate the coefficient of determination R² for LsqFit.jl result fit and measured data y.

RetentionData.collect_categories — Method

collect_categories(meta_data)

Collect the category entrys of the substances in the different datasets of meta_data, if they are available in a multidimensionsal array.

RetentionData.collect_csv_paths — Method

collect_csv_paths(folder)

Collect the paths of all .csv-files in the folder and all its sub-folders in an array of Strings.

RetentionData.dataframe_of_all — Method

datafram_of_all(meta_data::DataFrame)

Combine the separate dataframes with the parameter sets of the different entrys of the meta_data dataframe into one big dataframe.

RetentionData.duplicated_data — Method

duplicated_data(alldata)

Find the duplicated data entrys in alldata based on identical CAS-number and identical stationary phase. The dataframe alldata must not have missing CAS-number entrys.

Output

array of dataframe duplicated_data: every array element has a dataframe of the duplicated data

array of Bool duplicated_entry: true, if data has a duplicated, fals if not.

RetentionData.extract_meta_data — Method

extract_meta_data(csv_paths::Array{String,1})

Extract the meta data contained in the filename of a collection of paths to .csv-files. A dataframe with the columns:

path

filename

source

phase

beta0

Tref

d

gas

is returned. For the correct format of the filename see 'Note.md'.

RetentionData.extract_parameters_from_fit — Method

extract_paramaters_from_fit(fit, β0)

Extract the parameters A, B, C, Tchar, θchar and ΔCp from the fits of lnk over T data with the ABC-model and the Kcentric-model.

RetentionData.fit — Method

fit(model, T, lnk, lb, ub; weighted=false, threshold=NaN)

Old version. Fit the model-function to lnk over T data, with parameter lower boundaries lb and upper boundaries ub. Data points with a residuum above threshold are excluded and the fit is recalculated, until all residua are below the threshold or only three data points are left (unless threshold = NaN, which is default). If weighted = true the residuals of an unwheighted fit (OLS) are used as weights for a second fit ($w_i = 1/res_i^2$)

RetentionData.fit — Method

fit(model, T, lnk; check=true, ftrusted=0.7)

New version. Fit the model-function to lnk over T data. If check is true than a robust fit (RAFF.jl, parameter ftrusted) is executed and outliers are excluded from the fit with LsqFit.jl. If check is false than the outliers found with RAFF.jl are still used for the fit with LsqFit.jl.

Output

fit - the result from LsqFit.jl

robust - the result from RAFF.jl

outlier - the outliers found with RAFF.jl as a tuple of (T, lnk)

RetentionData.fit_models! — Method

fit_models!(meta_data::DataFrame, CheckBase; ftrusted=0.7)

New version. Fit the K-centric-model at the lnk(T) data of the data-array of dataframes, using RAFF.jl and LsqFit.jl and add the result in a new column (fit) of `meta_data

RetentionData.fit_models! — Method

fit_models!(meta_data::DataFrame; weighted=false, threshold=NaN, lb_ABC=[-Inf, 0.0, 0.0], ub_ABC=[0.0, Inf, 50.0], lb_Kcentric=[0.0, 0.0, 0.0], ub_Kcentric=[Inf, Inf, 500.0])

Old version. Fit the ABC-model and the K-centric-model at the lnk(T) data of the data-array of dataframes, using LsqFit.jl and add the result in a new column (fit) of `meta_data

RetentionData.fit_models — Method

fit_models(data::Array{DataFrame}, β0::Array{Float64})

New version. Fit the K-centric-model at the lnk(T) data of the data-array of dataframes, using LsqFit.jl

RetentionData.fit_models — Method

fit_models(data::Array{DataFrame}, β0::Array{Float64})

Old version. Fit the ABC-model and the K-centric-model at the lnk(T) data of the data-array of dataframes, using LsqFit.jl

RetentionData.flag — Method

flag(data)

Create a flag if certain conditions for the parameters are not fullfilled: - value of lambertw_x is < -1/e or > 0 - value of A > 0 - value of B < 0 - value of C < 0 - value of Tchar < -Tst - value of θchar < 0

RetentionData.flagged_data — Method

flagged_data(alldata::DataFrame)

Filter the parameter data for flagged and not-flagged datasets.

RetentionData.formula_to_dict — Method

formula_to_dict(formula)

Translate the formula string of a chemical substance into a dictionary, where the elements contained in the substance are the keys and the number of atoms are the values.

Example

julia> formula_to_dict("C14H20O")
Dict{String, Int64}("C" => 14, "H" => 20, "O" => 1)

RetentionData.indentify_parameters — Method

identify_parameters(data::Array{DataFrame})

Extract the column indices of the parameters and identify the type of parameters sets stored in the dataframes of data.

ABC-parameters:

column name containing "A" -> A

column name containing "B" -> B

column name containing "C_" or column name == "C" -> C

K-centric:

column name containing "Tchar" -> Tchar

column name containing "thetachar" -> thetachar

column name containing "DeltaCp" -> DeltaCp

thermodynamic parameters at reference temperature Tref

column name containing "DeltaHref" -> DeltaHref

column name containing "DeltaSref" -> DeltaSref

column name containing "DeltaCp" -> DeltaCp

additionally from meta_data a value for Tref is needed

Output

param_set: Array of Arrays of string, listening the types of parameter sets, possible values "ABC", "K-centric", "TD"

index: Array of dictionaries with the colum index of the parameters (parameters name is the key, the index is the value)

RetentionData.lambertw_x — Method

lambertw_x(A, B, C, KRef)

Calculate the argument for the Lambert W function (-1 branch) used to convert the ABC set to the K-centric set. The reference distribution coefficient KRef is a needed additional information (normally, it should be equal to the phase ratio β of the reference column, where the parameters where measured). The value should be -1/e ≤ x < 0.

RetentionData.load_allparameter_data — Method

load_allparameter_data(db_path)

Load the data files (.csv format) with the Parameters data from the folder db_path including all subfolders. Based on the loaded parameters, the parameters of the not included parameter sets are calculated. Additional information from the filenames are also saved.

Output

A dataframes with the following columns:

path: Path of the folder from where the data was loaded.

filename: Name of the file from where the data was loaded.

source: Short name for the source from where the data original is taken.

phase: Name of the stationary phase corresponding to the data.

beta0: The phase ratio corresponding to the data.

Tref: The reference temperature used for the thermodynamic parameter set. Optional parameter, if not available it has the value missing.

d: The column diameter. Optional parameter, if not available it has the value missing.

gas: The used gas for the mobile phase. Optional parameter, if not available it has the value missing.

data: Dataframes with the parameters of the three different parameter sets. See function all_parameters().

Note

For the naming convention of the filenames see Note.md.

RetentionData.load_csv_data! — Method

load_csv_data!(meta_data::DataFrame)

Load the data from the .csv-file located at meta_data.path, meta_data.filename. Data for DeltaHref is multiplied by 1000, if the corresponding unit is kJ/mol. The loaded data is returned in an additional column data in the meta_data dataframe.

RetentionData.load_csv_data — Method

load_csv_data(meta_data::DataFrame)

Load the data from the .csv-file located at meta_data.path, meta_data.filename. Data for DeltaHref is multiplied by 1000, if the corresponding unit is kJ/mol. The loaded data is returned in a dataframe. If an ID-list is available in the same folder the shortname/number will be looked up in the ID-list and replaced by the full name given in the ID-List.

RetentionData.load_custom_CI_database — Method

load_custom_CI_database(custom_database_url)

Load a custom database for ChemicalIdentifiers.jl from the location custom_database_url, if the custom database is not already loaded.

RetentionData.load_lnkT_data — Method

load_lnkT_data(db_path)

Load the data files (.csv format) with lnk-T data from the folder db_path including all subfolders.

RetentionData.load_parameter_data — Method

load_parameter_data(db_path)

Load the data files (.csv format) with the Parameters data from the folder db_path including all subfolders. Based on the loaded parameters, the parameters of the not included parameter sets are calculated. Additional information from the filenames are also saved.

Output

A dataframes with the following columns:

path: Path of the folder from where the data was loaded.

filename: Name of the file from where the data was loaded.

source: Short name for the source from where the data original is taken.

phase: Name of the stationary phase corresponding to the data.

beta0: The phase ratio corresponding to the data.

Tref: The reference temperature used for the thermodynamic parameter set. Optional parameter, if not available it has the value missing.

d: The column diameter. Optional parameter, if not available it has the value missing.

gas: The used gas for the mobile phase. Optional parameter, if not available it has the value missing.

data: Dataframes with the parameters of the three different parameter sets. See function all_parameters().

Note

For the naming convention of the filenames see Note.md.

RetentionData.new_database_format — Method

new_database_format(data; ParSet="Kcentric")

Export the data into the new database format with the columns:

Name

CAS

Phase

Tchar resp. A

thetachar resp. B

DeltaCp resp. C

phi0

Source

Cat ... several columns with categories

With the parameters ParSet="ABC" the ABC-parameters are exported, with ParSet="Kcentric" (default) the K-centric parameters are exported.

RetentionData.old_database_format — Method

old_database_format(data)

Export the data into the old database format with the columns:

Name

CAS

Cnumber

Hnumber

Onumber

Nnumber

Ringnumber

Molmass

Phase

Tchar

thetachar

DeltaCp

phi0

Annotation

RetentionData.plot_lnk_fit — Method

plot_lnk_fit(fit, i, j)

Plot the lnk-values over T of the selected dataset i of the selected substance j. Also the fit solution for the ABC-model and the K-centric-model are shown over the extended temperature range from 0°C to 400°C.

RetentionData.plot_res_lnk_fit — Method

plot_res_lnk_fit(fit, i, j)

Plot the absolute residuals (not the weighted) of lnk-values over T of the selected dataset i of the selected substance j to the fitted models.

RetentionData.ring_number — Method

ring_number(smiles)

Extract the number of rings of a substance defined by its SMILES. The highest digit contained in the SMILES is returned as the number of rings. Only single digit ring numbers are recognized.

RetentionData.save_all_parameter_data — Method

save_all_parameter_data(meta_data::DataFrame)

Save the data in the meta_data dataframe in new .csv-files in the same folder as the original data using the new filename Source_AllParam_Tablename_statPhase(_d)(_gas).csv. Source is the name of the source of the original data, statPhase is the stationary phase corresponding to the data and Tablename the name of the table of the original data. The optional entrys d and gas stand for the column diameter and the gas of the mobile phase.

RetentionData.substance_identification — Method

substance_identification(data::DataFrame)

Look up the substance name from the data dataframe with ChemicalIdentifiers.jl to find the CAS-number, the formula, the molecular weight MW and the smiles-identifier. If the name is not found in the database of ChemicalIdentifiers.jl a list with alternative names (shortnames.csv) is used. If there are still no matches, missing is used.